Switch to hono
While using raw `Deno.serve` works for now, we'd eventually outgrow it. Hono lets us keep things simple while also opening the door to other tooling to add future capabilites while continuing to keep the implementation small.
