+++ deno.json
@@ -1,6 +1,6 @@
 {
   "tasks": {
-    "dev": "deno run --allow-net --unstable-kv --watch main.ts"
+    "dev": "deno run --allow-net --unstable-kv --watch main.tsx"
   },
   "imports": {
     "@std/assert": "jsr:@std/assert@1",
@@ -10,6 +10,10 @@
     "project": "a49db880-a1d0-4f2b-9576-4e23cbaae630",
     "exclude": ["**/node_modules"],
     "include": [],
-    "entrypoint": "main.ts"
+    "entrypoint": "main.tsx"
+  },
+  "compilerOptions": {
+    "jsx": "precompile",
+    "jsxImportSource": "jsr:@hono/hono/jsx"
   }
 }
