+++ deno.lock
@@ -1,9 +1,13 @@
 {
   "version": "4",
   "specifiers": {
+    "jsr:@hono/hono@*": "4.6.8",
     "jsr:@std/ulid@*": "1.0.0"
   },
   "jsr": {
+    "@hono/hono@4.6.8": {
+      "integrity": "4c92cb895360d93b7bc637c0c3f26449761b2bb160009c0387a37a1d457850f0"
+    },
     "@std/ulid@1.0.0": {
       "integrity": "d41c3d27a907714413649fee864b7cde8d42ee68437d22b79d5de4f81d808780"
     }
